+++ config/bool.c
@@ -83,7 +83,7 @@ static int bool_string_set(const struct ConfigSet *cs, void *var, struct ConfigD
     return CSR_ERR_CODE; /* LCOV_EXCL_LINE */
 
   int num = -1;
-  for (unsigned int i = 0; i < mutt_array_size(bool_values); i++)
+  for (size_t i = 0; i < mutt_array_size(bool_values); i++)
   {
     if (mutt_str_strcasecmp(bool_values[i], value) == 0)
     {
